+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/home/config/model/HomeDataModel.java
@@ -1,12 +1,18 @@
 package io.github.sakurawald.fuji.module.initializer.home.config.model;
 
+import com.google.common.collect.HashBiMap;
 import com.google.gson.annotations.SerializedName;
-import io.github.sakurawald.fuji.module.initializer.home.structure.PlayerHomeMap;
+import io.github.sakurawald.fuji.core.structure.GlobalPos;
 import java.util.concurrent.ConcurrentHashMap;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 
+@Data
+@NoArgsConstructor
 public class HomeDataModel {
 
     @SerializedName(value = "name2home", alternate = "homes")
-    public @NotNull ConcurrentHashMap<String, PlayerHomeMap> name2home = new ConcurrentHashMap<>();
+    @NotNull ConcurrentHashMap<String, HashBiMap<String, GlobalPos>> name2home = new ConcurrentHashMap<>();
+
 }